@@ -1,6 +1,7 @@
 package com.fs.ru.registration.jpa
 
-import javax.persistence.*
+import jakarta.persistence.*
+import org.hibernate.Hibernate
 
 @Entity
 @Table(name = "roles")
@@ -14,4 +15,18 @@ data class Role(
     val name: String
 
 ) {
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
+        other as Role
+
+        return id == other.id
+    }
+
+    override fun hashCode(): Int = javaClass.hashCode()
+
+    @Override
+    override fun toString(): String {
+        return this::class.simpleName + "(id = $id , name = $name )"
+    }
 }
