@@ -1,6 +1,7 @@
 package com.fs.ru.registration.jpa
 
-import javax.persistence.*
+import jakarta.persistence.*
+import org.hibernate.Hibernate
 
 @Entity
 @Table(name = "users")
@@ -35,4 +36,19 @@ data class User(
         inverseJoinColumns = [JoinColumn(name = "role_id", referencedColumnName = "id")]
     )
     var roles: Collection<Role>? = null
-)
+) {
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
+        other as User
+
+        return id != null && id == other.id
+    }
+
+    override fun hashCode(): Int = javaClass.hashCode()
+
+    @Override
+    override fun toString(): String {
+        return this::class.simpleName + "(id = $id , username = $username , firstName = $firstName , lastName = $lastName , email = $email , password = $password , enabled = $enabled )"
+    }
+}
