@@ -1,9 +1,10 @@
 package com.fs.ru.registration.jpa
 
+import jakarta.persistence.*
+import org.hibernate.Hibernate
 import java.sql.Date
 import java.sql.Timestamp
 import java.util.*
-import javax.persistence.*
 
 @Entity
 @Table(name = "verification_token")
@@ -24,6 +25,21 @@ data class VerificationToken(
     val user: User
 ) {
     constructor(token: String?, user: User) : this(0, token, calculateExpiryDate(1440), user)
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
+        other as VerificationToken
+
+        return id != null && id == other.id
+    }
+
+    override fun hashCode(): Int = javaClass.hashCode()
+
+    @Override
+    override fun toString(): String {
+        return this::class.simpleName + "(id = $id , token = $token , expiryDate = $expiryDate , user = $user )"
+    }
 }
 
 private fun calculateExpiryDate(expiryTimeInMinutes: Int): Date {
