@@ -1,13 +1,14 @@
 package com.fs.ru.registration.jwt
 
+import jakarta.servlet.ServletException
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 import org.slf4j.LoggerFactory
 import org.springframework.security.core.AuthenticationException
 import org.springframework.security.web.AuthenticationEntryPoint
 import org.springframework.stereotype.Component
 import java.io.IOException
-import javax.servlet.ServletException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+
 
 @Component
 class JwtAuthEntryPoint : AuthenticationEntryPoint {
@@ -26,4 +27,6 @@ class JwtAuthEntryPoint : AuthenticationEntryPoint {
     companion object {
         private val logger = LoggerFactory.getLogger(JwtAuthEntryPoint::class.java)
     }
+
+
 }
