@@ -1,10 +1,10 @@
 package com.fs.ru.registration.repository
 
 import com.fs.ru.registration.jpa.User
+import jakarta.transaction.Transactional
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.repository.query.Param
 import java.util.*
-import javax.transaction.Transactional
 
 interface UserRepository : JpaRepository<User, Long> {
 
